Deprecated parallax, renamed to background
