+++ examples/package.json
@@ -9,6 +9,7 @@
     "ory-editor-plugins-html5-video": "0.0.0",
     "ory-editor-plugins-image": "0.0.0",
     "ory-editor-plugins-parallax-background": "0.0.0",
+    "ory-editor-plugins-background": "0.0.0",
     "ory-editor-plugins-slate": "0.0.0",
     "ory-editor-plugins-spacer": "0.0.0",
     "ory-editor-plugins-video": "0.0.0",
