+++ packages/plugins/layout/background/src/components/Color.js
