+++ packages/plugins/layout/background/src/components/Image.js
