+++ packages/plugins/layout/background/src/components/LinearGradient.js
