+++ packages/plugins/layout/background/src/index.css
@@ -20,11 +20,18 @@
  *
  */
 
-.ory-plugins-layout-parallax-background {
-  background-attachment: fixed;
+.ory-plugins-layout-background {
   background-position: center;
   background-repeat: no-repeat;
   background-size: cover;
   color: white;
   padding: 12px;
+}
+
+.ory-plugins-layout-background__backstretch {
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
 }
